requirements.txt
asgiref==3.8.1
Django==4.1.13
django-taggit==3.0.0
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.3
